Day09 The Complete IntCode Computer
Day09 Add the relative mode, migrate from int to long

Day09 Migrate from long to int, add some checks

Day09 Test quine

Day09 Remove the Long range checking, revert change and fix 07

Day09 Allow to decide if should pause on output and fix the quine

Day09 Revert the overflow checking

Day09 Simplify day 05

Day09 WIP fix instruction at 50th position

Day09 Fix error 203 at position 50

      TODO:
      * fix error 21108 at position 405
      * fix error 21101 at position 570

Day09 Fix the errors and solve the 1st part

Day09 Clean up

Day09 Solve the second part

+++ src/main/java/year2019/Day02.java
@@ -14,7 +14,7 @@ public class Day02 implements IAocTask {
 
     @Override
     public void solvePartOne(List<String> lines) {
-        int[] parsedCode = loadProgram(lines);
+        long[] parsedCode = loadProgram(lines);
         parsedCode[1] = 12;
         parsedCode[2] = 2;
         runProgram(parsedCode);
@@ -23,23 +23,23 @@ public class Day02 implements IAocTask {
 
     @Override
     public void solvePartTwo(List<String> lines) {
-        int[] parsedCode = loadProgram(lines);
-        int[] copied = runProgramV2(parsedCode);
+        long[] parsedCode = loadProgram(lines);
+        long[] copied = runProgramV2(parsedCode);
         System.out.format("%d", 100 * copied[1] + copied[2]);
     }
 
-    private void runProgram(int[] parsedCode) {
+    private void runProgram(long[] parsedCode) {
         int i = 0;
         for (; i < parsedCode.length; ) {
             i = runBasicInstructions(parsedCode, i);
         }
     }
 
-    private int[] runProgramV2(int[] parsedCode) {
+    private long[] runProgramV2(long[] parsedCode) {
 
         for (int i = 0; i < 100; i++) {
             for (int j = 0; j < 100; j++) {
-                int[] copied = new int[parsedCode.length];
+                long[] copied = new long[parsedCode.length];
                 System.arraycopy(parsedCode, 0, copied, 0, parsedCode.length);
                 copied[1] = i;
                 copied[2] = j;
@@ -53,7 +53,7 @@ public class Day02 implements IAocTask {
         return parsedCode;
     }
 
-    public static int runBasicInstructions(int[] parsedCode, int i) {
+    public static int runBasicInstructions(long[] parsedCode, int i) {
         if (parsedCode[i] == INSTR_STOP) {
             i = parsedCode.length;
         } else if (parsedCode[i] == INSTR_ADD) {
@@ -66,11 +66,11 @@ public class Day02 implements IAocTask {
         return i;
     }
 
-    public static void addNumbers(int i, int[] parsedCode) {
-        parsedCode[parsedCode[i + 3]] = parsedCode[parsedCode[i + 1]] + parsedCode[parsedCode[i + 2]];
+    public static void addNumbers(int i, long[] parsedCode) {
+        parsedCode[(int) parsedCode[i + 3]] = parsedCode[(int) parsedCode[i + 1]] + parsedCode[(int) parsedCode[i + 2]];
     }
 
-    public static void multiplyNumbers(int i, int[] parsedCode) {
-        parsedCode[parsedCode[i + 3]] = parsedCode[parsedCode[i + 1]] * parsedCode[parsedCode[i + 2]];
+    public static void multiplyNumbers(int i, long[] parsedCode) {
+        parsedCode[(int) parsedCode[i + 3]] = parsedCode[(int) parsedCode[i + 1]] * parsedCode[(int) parsedCode[i + 2]];
     }
 }
